protect routes, auth and guest middleware default

+++ routes/web.php
@@ -19,31 +19,31 @@ use App\Models\Listing;
 
 Route::get('/', [ListingController::class, 'index']);
 
-Route::get('/listings/create', [ListingController::class, 'create']); // Show Create Form
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth'); // Show Create Form
 
-Route::post('/listings', [ListingController::class, 'store']); // Store
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth'); // Store
 
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']); // Show Edit Form
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth'); // Show Edit Form
 
-Route::put('/listings/{listing}', [ListingController::class, 'update']); // Edit
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth'); // Edit
 
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']); // Delete
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth'); // Delete
 
 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 
 // SHOW REGISTER CREATE FORM
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // STORE NEW USER
 Route::post('/users', [UserController::class, 'store']);
 
 // LOG OUT
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // SHOW LOG IN FORM
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // LOG IN
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
